pyproject.toml
[project]
name = "oauth-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "msal>=1.34.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
